requirements.txt
loguru
mysql-connector-python
flask
